refactor(auth): extract logic from CredentialsSignInForm into container component
- Moved `CredentialsSignInForm` to `@/containers/auth/credentials-sign-in-form.tsx`
- Created `SignInForm` presentation component at `@/components/auth/sign-in/sign-in-form.tsx`
- Updated imports in `@/app/(auth)/acesso/entrar/page.tsx` to use the new container path
